% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpo.R
\name{cpo_fc}
\alias{cpo_fc}
\title{CPO - Conditional Predictive Ordinate - for srplv}
\usage{
cpo_fc(matrix.cpo)
}
\arguments{
\item{matrix.cpo}{Matrix with dimension n.sys by n.size containing CPO values for n.sys systems and n.size posterior sample. Result from \code{\link{MWG}}.}
}
\value{
sum of log of CPO.
}
\description{
\code{cpo_fc} calculates the sum of logarithm of CPO.
}
