% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{.gettime}
\alias{.gettime}
\title{GetTime
This function returns the current amount of spent time by the current R process (user time + system time). The unit is in minutes.}
\usage{
.gettime()
}
\description{
GetTime
This function returns the current amount of spent time by the current R process (user time + system time). The unit is in minutes.
}
