% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_alg.R
\name{EMalg}
\alias{EMalg}
\title{EM algorithm for srplv}
\usage{
EMalg(guess, Data, distribution, L)
}
\arguments{
\item{guess}{The initial values for the parameters. The default is (1,1).}

\item{Data}{Data in the srplv format as \code{\link{data_to_srplv}}.}

\item{distribution}{The chose distribution: weibull, gamma, lnorm, llogis.}

\item{L}{Number of Monte Carlo simulation.}
}
\value{
A list with the following components:
\item{values}{The values of parameters in all EM algorithm iterations.}
\item{est.par}{The maximum likelihood estimates of the parameters.}
\item{max.logLik}{The maximum likelihood value.}
\item{n.it}{The number of EM iterations.}
\item{time}{The computational time.}
}
\description{
\code{EMalg} returns the maximum likelihood estimates obtained by EM algorithm.
}
