% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Data generation of Superposed Renewal Processes}
\usage{
generate_data(n.sys, n.comp, m.t, v.t, m.c, v.c, distribution, s.seed)
}
\arguments{
\item{n.sys}{The number of systems.}

\item{n.comp}{The number of sockets for each system.}

\item{m.t}{The expected failure time of components.}

\item{v.t}{The variance of failure time of components .}

\item{m.c}{The expected failure time of censor (end-of-observation time).}

\item{v.c}{The variance of failure time of censor (end-of-observation time).}

\item{distribution}{The chose distribution: weibull, gamma, lnorm, llogis.}

\item{s.seed}{The seed for data generation.}
}
\value{
A list with the following components:
\item{n}{the number of systems in the fleet.}
\item{n.comp}{the number of sockets in each system.}
\item{r}{a vector with a length n, recording the number of failures for each system.}
\item{time}{a list that each vector records the failure times of each system.}
\item{cens}{a vector with a length n, recording the end-of-observation time for each system.}
\item{par.t}{a vector with parameters of components failure times distribution.}
\item{par.c}{a vector with parameters of censor distribution.}
\item{delta.true}{a list that each vector records the socket index failure indicator of each system.}
}
\description{
\code{generate_data} generates srplv dataset.
}
