% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\name{initials}
\alias{initials}
\title{Initial values for the parameters}
\usage{
initials(Data, distribution)
}
\arguments{
\item{Data}{Data in the srplv format as \code{\link{data_to_srplv}}.}

\item{distribution}{The chose distribution: weibull, gamma, lnorm, llogis.}
}
\value{
A vector with initial parameters values.
}
\description{
\code{initials}  obtains initial values for the parameters.
}
