% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{lik.par}
\alias{lik.par}
\title{This function calculates the mean of likelihood function evaluated at the vector par over L values.}
\usage{
lik.par(par, Data, delta, L, distribution)
}
\description{
This function calculates the mean of likelihood function evaluated at the vector par over L values.
}
