% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_functions.R
\name{lik}
\alias{lik}
\title{Function lik calculates log likelihood contribution of the fleet of systems.}
\usage{
lik(par, Data, delta, distribution)
}
\description{
Function lik calculates log likelihood contribution of the fleet of systems.
}
