% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_functions.R
\name{liksyst_weibull}
\alias{liksyst_weibull}
\title{Function liksyst calculates log likelihood contribution of a single system.}
\usage{
liksyst_weibull(par, r, time, d, m, cens)
}
\description{
Function liksyst calculates log likelihood contribution of a single system.
}
