% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures_loglik.R
\name{measures_loglik}
\alias{measures_loglik}
\title{Selection criteria based on maximized log-likelihood function value}
\usage{
measures_loglik(n, npar, loglik)
}
\arguments{
\item{n}{Number of systems.}

\item{npar}{Number of the parameters.}

\item{loglik}{Maximized log-likelihood function value returned in \code{\link{EMalg}}.}
}
\value{
A list with the following components:
\item{loglik}{The maximized log-likelihood function value.}
\item{aic}{AIC (The Akaike Information Criterion).}
\item{aicc}{AICc (The corrected Akaike Information Criterion).}
\item{bic}{BIC (The Bayesian Information Criterion).}
\item{hqic}{HQIC (The Hannan-Quinn Information Criterion).}
\item{caic}{CAIC (The Consistent Akaike Information Criterion).}
}
\description{
\code{measures_loglik} calculates the AIC, AICc, BIC, HQIC, CAIC measures.
}
