% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWG.R
\name{MWG}
\alias{MWG}
\title{Metropolis within Gibbs algorithm for srplv}
\usage{
MWG(guess, Data, distribution, burn, jump, prior1, prior2, n.size = 1000)
}
\arguments{
\item{guess}{The vector of initial values for the parameters. The default is (1,1).}

\item{Data}{Data in the srplv format as \code{\link{data_to_srplv}}.}

\item{distribution}{The chose distribution: weibull, gamma, lnorm, llogis.}

\item{burn}{Burn-in sample.}

\item{jump}{Jump sample.}

\item{prior1}{Hyperparameters for the prior distribution for parameter 1.}

\item{prior2}{Hyperparameters for the prior distribution for parameter 2.}

\item{n.size}{Number of MCMC sample. The default is 1000.}
}
\value{
A list with the following components:
\item{post.values}{The sample from the posterior distribution for each parameter (column) with size n.size.}
\item{acceptance}{The acceptance rate.}
\item{time}{The computational time.}
\item{matrix.cpo}{Matrix with dimension n.sys by n.size containing Conditional Predictive Ordinate (CPO) values for n.sys systems and n.size posterior sample.}
}
\description{
\code{MWG}  obtains MCMC sample from posterior distribution via MWG algorithm for  superposed renewal processes by means of latent variables.
}
