% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{post.weibull}
\alias{post.weibull}
\title{Function posteriori calculates the kernel of posterior density.}
\usage{
post.weibull(par, Data, delta, distribution, prior1, prior2)
}
\description{
Function posteriori calculates the kernel of posterior density.
}
