% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_latent_var.R
\name{rdelta}
\alias{rdelta}
\title{Latent variables generation}
\usage{
rdelta(par, Data, distribution)
}
\arguments{
\item{par}{The vector of parameters.}

\item{Data}{Data in the srplv format as \code{\link{data_to_srplv}}.}

\item{distribution}{The chose distribution: weibull, gamma, lnorm, llogis.}
}
\value{
A list with the latent variables for all systems in the fleet.
}
\description{
\code{rdelta} generates latent variables for all systems in the fleet.
}
